Book: removed repeated paragraph in Invoking NoiseChisel section
Until now, the ending of the Invoking NoiseChisel section in the book was
repeated in two paragraphs (and one of them was from the time that
NoiseChisel and Segment were not yet separated).

With this commit the repeated old paragraph has been removed. Also, an
extra "here1" that was printed with 'astscript-color-faint-gray' has been
removed.

The problem in the book was reported by Hamed Altafi.