ConvertType: standard input only read when no file specified
Until now, ConvertType would always read the standard input; even when the
three or four color channels had been given as files on the command-line!
This is contrary to normal expection on the command-line (for example AWK
or SED only check the standard input when no file is given as an
argument). Therefore it could cause un-expected bugs in situations where
ConvertType was being called within a loop that was parsing the standard
input (see [1]).

With this commit, ConvertType now only reads the standard input when no
input file is specified.

This issue was reported by Jesús Vega and Raul Infante-Sainz and solved
with the help of Greg Wooledge and Dennis Williamson.