Makefile extensions: new function to return previous element in list
Until now, when using Gnuastro within a large pipeline, it was not easy to
serialize only one part of a pipeline in Makeafiles! Make's '.NOTPARALLEL'
target would also serialize all the prerequisites of the given targets.

With this commit, a new Makefile extension function has been added to
Gnuastro for this purpose: 'ast-text-prev-in-list'. It will return the
previous word to the given word and when placed as a prerequisite, it will
serialize the operations.