Arithmetic: --append adds output to new HDU of existing output
Until now, when you wanted to write the output of Arithmetic as a HDU of an
already existing FITS file, you had to make a temporary file, then use the
Fits program to add the HDU to that file. This would make your pipeline
code complex and be slow. Previously in the Crop program, we had a similar
problem, and as a solution, we added the '--append' option.

With this commit, this option is also implemented in Arithmetic.