color-faint-gray: warning the user to convert zero pixel values to nan
Until now, we did not inform the user to convert all pixels that are equal
to zero to nan values. We are used to having such pixels already as NaNs
(from Gnuastro programs). However, there are many images coming from other
software that put zero values to those pixels, normally in the
corner/borders, or in general in the outer part of the images. These zero
pixel values that are meaningless affect the estimation of the necessary
parameters for constructing a nice color/gray image.

With this commit, a couple of sentences explaining this situation have been
included to warn the user. The reporting of the default/used values for
grayval and colorval was wrong and it has been corrected. In addition to
that, I also fixed a couple of typos in other part of the Book.