MakeCatalog: correct command in suggested statement of error message
Until now, when MakeCatalog was given a labels image with negative values,
it would suggest an Arithmetic command to correct the input in case the
input was from a clumps image. However, the command had an extra 'i' in it;
causing the Arithmetic command to produce an error.

With this commit, the suggested command has been amended to remove any 'i's
(which were redundant for this suggestion; but necessary in the previous
proposed command).

This bug was reported by Zahra Sharbaf.

This fixes bug #65377.