astscript-radial-profile: dealing with all possible azimuth angle values
Until now, when providing azimuth angles that were not in the range 0-360
deg, the script provided wrong results. For example, when providing
--azimuth=337,367, it creates the wedgle between 337-360 degrees. So, it
stops the wedgle at 360 degrees. Same would happen with negative or
arbitrary large angle values.

With this commit, a function is added to correct the input azimuthal range
in case it is outside the range of 0 to 360 degrees before any further
azimuthal angle work is done.

This bug was reported by Fernando Buitrago Alonso.

This fixes bug #65382.