Statistics: does not crash when all inputs are NaN
Until now, when input data for the Statistics program is all NAN, it would
abort with an error of not having any data. This is not a desired behaviour
because complicates processing within complex pipelines. When there is no
input, the output should be NaN.

With this commit, an all NAN (or empty) input will no longer result in an
error, but just shows a warning which the user can avoid by using the
'--quiet' option. Any resulting statistic (like '--median' or '--std') will
also be returned as a NaN.

This bug was reported by Raul Infante-Sainz.

This fixes bug #63064.