Statistics: MAD only calculated in one-row outputs with --mad
Until now, if the user requested '--median' or '--mad', the statistics
program would call 'gal_statistics_median_mad' (which calculates both!). As
a result, if the user just wanted the median (and not the MAD), the program
would un-necessarily become slow (to calculate the un-used MAD!).

With this commit, if the MAD is not requested, the Statistics program will
just call 'gal_statistics_median' to only do the requested job and nothing
extra.