Library (fits.h): older CFITSIO versions correctly stored
Until now, when Gnuastro was compiled with older versions of CFITSIO (which
had a three component version string), a 'char *' was passed to 'asprintf'
instead of a 'char **'. This would cause a segmentation fault when Gnuastro
was compiled with the default optimization.

With this commit, the correct pointer is passed and the segmentation fault
is fixed.

This bug was reported by Giacomo Lorenzetti.

This fixes bug #65494.