Tests: missing check_with_program (for valgrind) commands fixed
Until now, some of the tests were missing the '$check_with_program' command
at the start of the final command to run the test. Currently this is used
for optionally running the test within Valgrind.

With this commit, this command has been placed at the start of those tests
that didn't have it. In particular, it was the tests on the 3D features of
NoiseChisel, Segment and MakeCatalog (that were developed on a separate
branch for a long time; and I forgot to bring this feature into them during
the merge).