Crop: returned value to shell is non-zero if no outputs made
Until now, when the Crop program was given a crop that did not overlap with
the input image(s), it would not produce any output, and return control
back to the shell successfully! However, this is not expected behavior in
the command-line and many shell scripts or pipelines depend on the return
value for the evaluation of their next stages.

With this commit, when no outputs are created by crop, it will return a
value of 0; allowing easy usage of the shell's conditionals to adjust
automatic pipelines.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #65571.