Warp: avoid pixels affeced by projections to infinity
Until now, when a projection to infinity was met, the program generated an
undefined behaviour due to a division by zero, also increasing the
execution time (due to the very large range of pixels that needed to be
checked unnecessarily).

With this commit, points projected to infinity are treated as NaNs.  Those
are the pixels whose corners are on the slope T[6]*x+T[7]*y+T[8]=0, where T
is the matrix (in the detected case, the inverse of the warping matrix) and
x and y are the coordinates of each pixel's corners (output, in the case).
In order to avoid floating point issues, zero is "defined" as the range
between '-DBL_EPSILON' and 'DBL_EPSILON'.

The cause of the bug was detected using the '-fsanitize' flags
'float-cast-overflow' and 'float-divide-by-zero'.

This fixes bug #65561.