Makefile extensions: new function to batch by available RAM
Until now, in the Makefile we only had 'makeplugin_text_prev_batch_in_list'
function which required you to manually insert the number of files in each
batch. But when we go from one system to another, depending on the
available RAM and number of threads, that number needs to be changed. This
would make the pipeline non-automatic, which is very bad and potentially
causes many bugs.

Furthermore the underlying algorithm that was used to divide the input
files into batches was not very optimal (would parse the full list for
every target, do many extra allocations and etc).

With this commit, a new function has been added that will read the
avialable RAM in the operating system and take the amount of RAM needed by
each instance of the recipe to automatically find the number of elements in
each batch.

Furthermore, a new algorithm is now used to divide the input files into
batches: when Make is run in parallel on many files, this is much faster
(by almost 50% for 12 threads) than the previous algorithm. The new
function is now also used in the 'makeplugin_text_prev_batch_in_list'
(which has now lost the redundant '_in_list' suffix to make it easier to
use).