Library (arithmetic): correctly accounting for all-NaN pixels
Until now, when a pixel was NaN in all the inputs of the stacking
operators, they would crash with a segmentation fault because of the newly
changed feature in the statistics operators (that would free the input
array's pointer and was necessary for many low-level checks).

With this commit, before giving the pixel value in all the inputs to the
respective statistics operator, we count how many are not blank and only
call the statisitics operators when we actually have data.