Makefile extensions: batch by RAM takes empty RAM usage
Until now, when calling the newly added 'ast-text-prev-batch-by-ram'
function, giving a RAM value (as a number in Gigabytes) was
mandatory. However, when this function is used in a Makefile that is itself
later used in higher-level script, it happens that the RAM check is only
necessary in some calls to the Makefile, not all. In such cases, the user
would have to put the whole target+prerequisite structure in conditionals,
making the Makefile hard to read and buggy.

With this commit, as in other GNU Make functions, this Gnuastro Makefile
extension now returns 'NULL' when the value in the gigabyte argument is an
empty string (removing its effect in the prerequisite list).