Makefile extensions: sanity checks added to ast-text-prev functions
Until now, when any of the 'ast-text-prev*' functions were given an empty
string for their argument, they would produce a segmentation fault (because
they were reading from a non-allocated string!).

With this commit, a sanity check function has been added at the start of
all Gnuastro's Makefile extension functions. It will return an empty string
when any of the input arguments is an empty string.